Create logic6.cs
You are signing up for a new email address. 
Write a programme that checks whether the email address you would like is available.
You will need to create a variable containing an email address which is already taken.
If the email address is available, produce a message that says “Yes variable is available for you”.
If the email address is not available, produce a message that says “Sorry variable is being used by another user”.
Paste both outputs underneath your code.